requirements.txt
Django==5.2.5
PyMuPDF==1.24.9
scikit-learn==1.5.2
Pillow==11.3.0
